Postproc v 0.2
Here we have rallpack3 that produces almost perfect results. After
changes to current injection in steps 3 and occupancy in steps 4.

Rallpack 3 works. The other scripts could be outdated.

In Rallpack 3 we have the refined data that is injected in the ks tests
to be used as raw data for the pvalue statistics. More in the
rallpack 3 script itself.

Refined data (parquet format) are added so that people can jump start on
plotting the statistics they prefer.

Polished after Iain asked for data to play around in the rallpack 3
confluence page the 4/11/2021