low ram footprint (#18)
Remove `raw_traces` caching and rework plots

### Raw traces:

**Before:**
- all raw traces are loaded
- all traces are refined

**Now:**
For each trace:
  - Trace is loaded
  - Trace is refined
  - the raw trace is dropped from memory if not needed anymore

This allows us to deal with Gbs of traces in a serial way. The
disadvantages are that some small functionalities are not available
in this way. For example, we cannot plot the raw_traces anymore.

### Plots:

**Taking care of reviewer 2**:

Rallpack1, 3, and Caburst mains were split in smaller functions
for readability. In addition multiple small changes were introduced
to improve plot readability. 

- Now figures require more fine tuning but are much more flexible
than before for adding stuff.
- additional goodness of fit methods

**Rallpack1**:
- add inset in plot
- Fixup! rallpack1 with same dt as analytic solution

**Rallpack3**:
- high channel density2 is the standard now
- add inset in plot
- readd batched results
- Fixup! range selection for `peaks_y` and `peaks_t`


Fix #17 

Authored-by: Alessandro Cattabiani <alessandro.cattabiani@epfl.ch>